@@ -1,9 +1,8 @@
-﻿namespace GildedRoseKata
+﻿namespace GildedRoseKata;
+
+public class Item
 {
-    public class Item
-    {
-        public string Name { get; set; }
-        public int SellIn { get; set; }
-        public int Quality { get; set; }
-    }
+    public string Name { get; set; }
+    public int SellIn { get; set; }
+    public int Quality { get; set; }
 }
