@@ -1,18 +1,15 @@
-﻿using Xunit;
-using System.Collections.Generic;
-using GildedRoseKata;
+﻿using GildedRoseKata;
 
-namespace GildedRoseTests
+namespace GildedRoseTests;
+
+public class GildedRoseTest
 {
-    public class GildedRoseTest
+    [Fact]
+    public void Foo()
     {
-        [Fact]
-        public void foo()
-        {
-            IList<Item> Items = new List<Item> { new Item { Name = "foo", SellIn = 0, Quality = 0 } };
-            GildedRose app = new GildedRose(Items);
-            app.UpdateQuality();
-            Assert.Equal("fixme", Items[0].Name);
-        }
+        List<Item> items = [ new Item { Name = "foo", SellIn = 0, Quality = 0 } ];
+        GildedRose app = new(items);
+        app.UpdateQuality();
+        Assert.Equal("fixme", items[0].Name);
     }
 }
