@@ -1,30 +1,20 @@
-﻿
+﻿using System.Text;
 using GildedRoseKata;
 
-using System;
-using System.IO;
-using System.Text;
-using System.Threading.Tasks;
+namespace GildedRoseTests;
 
-using VerifyXunit;
-
-using Xunit;
-
-namespace GildedRoseTests
+public class ApprovalTest
 {
-    public class ApprovalTest
+    [Fact]
+    public Task ThirtyDays()
     {
-        [Fact]
-        public Task ThirtyDays()
-        {
-            var fakeoutput = new StringBuilder();
-            Console.SetOut(new StringWriter(fakeoutput));
-            Console.SetIn(new StringReader("a\n"));
+        var fakeOutput = new StringBuilder();
+        Console.SetOut(new StringWriter(fakeOutput));
+        Console.SetIn(new StringReader("a\n"));
 
-            Program.Main(new string[] { "30" });
-            var output = fakeoutput.ToString();
+        Program.Main(["30"]);
+        var output = fakeOutput.ToString();
 
-            return Verifier.Verify(output);
-        }
+        return Verifier.Verify(output);
     }
 }
